GenomeAnalyzer/requirements_genome_analyzer.txt
biopython>=1.79
requests>=2.28.0

